Fix linking issues on 32-bit Windows
Since clock compiles its bundled C code with -O3, gcc will optimize
a division followed by a modulo calculation to a single divmod.
This requires a library call found in libgcc, so it is necessary to
link against it. (Not doing so caused #50.)

For more information, see the discussion at
https://ghc.haskell.org/trac/ghc/ticket/15094.