fix Num multiplication, add Real, Enum, Integral instances
These instances are based on using the nanosecond as the underlying unit,
i.e. the value of a TimeSpec is the integer s * 1e9 + ns.
Before it was based on representing the rational s + (ns / 1e9).
But the integer valuation satisfies the ring
laws and provides much more intuitive behavior, e.g.
(2 :: TimeSpec)*(2 :: TimeSpec) = (4 :: TimeSpec)
Before, this gave 0.